Renaming folder to containers and tweaking the code to match with the requirements
