+++ src/containers/App.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import BookList from './bookslist';
-import BooksForm from './booksform';
+import BookForm from './bookform';
 
 const App = () => (
   <div>
     <BookList />
-    <BooksForm />
+    <BookForm />
   </div>
 );
