+++ src/reducers/categories.js
@@ -2,7 +2,6 @@ import { SET_CATEGORY } from '../actions/index';
 import { FILTERS } from '../constants/constants';
 
 const category = (state = FILTERS.all, action) => {
-  console.log(action.payload);
   switch (action.type) {
     case SET_CATEGORY:
       return action.payload.filter;
